% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.numeric.waldCI.R
\name{as.numeric}
\alias{as.numeric}
\title{Convert waldCI to numeric}
\usage{
as.numeric(x, ...)
}
\arguments{
\item{x}{waldCI object}

\item{...}{ignored}
}
\value{
Length 2 vector of the bounds of `x`.
}
\description{
Convert waldCI to numeric
}
