% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{level}
\alias{level}
\alias{level,waldCI-method}
\alias{mean.waldCI}
\alias{sterr}
\alias{sterr,waldCI-method}
\alias{lb}
\alias{lb,waldCI-method}
\alias{ub}
\alias{ub,waldCI-method}
\alias{level<-}
\alias{level<-,waldCI-method}
\alias{mean<-}
\alias{mean<-,waldCI-method}
\alias{sterr<-}
\alias{sterr<-,waldCI-method}
\alias{lb<-}
\alias{lb<-,waldCI-method}
\alias{ub<-}
\alias{ub<-,waldCI-method}
\title{Getters and Setters for Wald CI objects}
\usage{
level(object)

\S4method{level}{waldCI}(object)

\method{mean}{waldCI}(x, ...)

sterr(object)

\S4method{sterr}{waldCI}(object)

lb(object)

\S4method{lb}{waldCI}(object)

ub(object)

\S4method{ub}{waldCI}(object)

level(object) <- value

\S4method{level}{waldCI}(object) <- value

mean(object) <- value

\S4method{mean}{waldCI}(object) <- value

sterr(object) <- value

\S4method{sterr}{waldCI}(object) <- value

lb(object) <- value

\S4method{lb}{waldCI}(object) <- value

ub(object) <- value

\S4method{ub}{waldCI}(object) <- value
}
\arguments{
\item{object}{A `waldCI` object}

\item{x}{A `waldCI` object}

\item{...}{ignored}

\item{value}{New value when setting}
}
\value{
Getters return scalar; setters return the object
}
\description{
Getters and Setters for Wald CI objects
}
