% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{contains}
\alias{contains}
\alias{overlap}
\title{Check whether a CI overlaps with a constant (`contains`) or another CI
(`overlap`)}
\usage{
contains(ci, value)

overlap(ci1, ci2)
}
\arguments{
\item{ci}{A `waldCI` object}

\item{value}{A numeric value}

\item{ci1}{A `waldCI` object}

\item{ci2}{A `waldCI` object}
}
\value{
Logical.
}
\description{
Check whether a CI overlaps with a constant (`contains`) or another CI
(`overlap`)
}
