% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldCI.R
\name{.getBounds}
\alias{.getBounds}
\title{Internal function to compute bounds from stored mean/sterr}
\usage{
.getBounds(ci)
}
\arguments{
\item{ci}{Confidence level}
}
\value{
A vector of length 2 for lower and upper bounds
}
\description{
Internal function to compute bounds from stored mean/sterr
}
