% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{.getSlot}
\alias{.getSlot}
\title{Internal function to get a named slot}
\usage{
.getSlot(object, slotname)
}
\arguments{
\item{object}{A `waldCI` object}

\item{slotname}{Which slot? Character.}
}
\value{
A length-1 vector with the requested `slotname`.
}
\description{
Internal function to get a named slot
}
