% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldCI.R
\name{makeCI}
\alias{makeCI}
\title{Create `waldCI` object. Either `lb` and `ub`, or `mean` and `sterr` must be
provided.}
\usage{
makeCI(level, lb = NULL, ub = NULL, mean = NULL, sterr = NULL)
}
\arguments{
\item{level}{Confidence level}

\item{lb}{lower bound, optional}

\item{ub}{upper bound, optional}

\item{mean}{mean, optional}

\item{sterr}{standard error, optional}
}
\value{
A `waldCI` object
}
\description{
Create `waldCI` object. Either `lb` and `ub`, or `mean` and `sterr` must be
provided.
}
\examples{
 myci <- makeCI(.95, mean = 2, sterr = 1)
}
