% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldCI.R
\name{show,waldCI-method}
\alias{show,waldCI-method}
\title{Show method for waldCI object}
\usage{
\S4method{show}{waldCI}(object)
}
\arguments{
\item{object}{A waldCI object}
}
\value{
`object`, invisibly
}
\description{
Show method for waldCI object
}
