% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCI.R
\name{transformCI}
\alias{transformCI}
\title{Apply a monotonic transformation to a `waldCI`}
\usage{
transformCI(ci, fn)
}
\arguments{
\item{ci}{A `waldCI` object}

\item{fn}{Any function that takes in a scalar and returns a scalar.}
}
\value{
The transformed waldCI
}
\description{
Monotonic functions are *not* enforced, but non-monotonic functions that
return nonsense results (e.g. ub < lb) will likely error.
}
